Move first query service to ./queries
