+++ packages/hub/node-tests/routes/merchant-infos-test.ts
@@ -461,7 +461,7 @@ describe('GET /api/merchant-infos', function () {
 
   it('fetches merchant infos available for association to a new card space', async function () {
     let cardSpaceQueries = await getContainer().lookup('card-space-queries');
-    let merchantInfoQueries = await getContainer().lookup('merchant-info-queries');
+    let merchantInfoQueries = await getContainer().lookup('merchant-info', { type: 'query' });
 
     // 3 merchants (jerry, kramer, george) belonging to first user
     let merchantInfoId = 'c5cd6479-ec74-4ecd-9aa6-96bdb02d255e';
